// src/lib/types.ts
export interface Equipment {
  tag: string;
  name: string;
  type: 'Building' | 'Crane' | 'Tank' | 'Terrain' | 'Pipe' | 'Valve';
  sistema?: string;
  area?: string;
  operationalState?: string;
  product?: string;
  position: { x: number; y: number; z: number };
  rotation?: { x: number; y: number; z: number };
  size?: { width: number; height: number; depth: number };
  radius?: number;
  height?: number;
  color: string;
  details?: string;
}
// ...
export type ColorMode = 'Produto' | 'Estado Operacional' | 'Equipamento';